Add work on pulling Eping notifications direct
Adds a script able to pull and filter out Eping notifications directly,
rather than having to rely on the painful system of email notifications.

This relies on JSON API provided implicitly by the web client. In an
ideal world we would use the search filters provdided server-side. By
frankley, I don't trust them and don't want to reverse engineer them.

Instead we know that there tends to be <100 notifications a day,
bandwidth is cheap and filtering locally is fast and controllable. So
lets stick with that.

Also if your from the WTO technical barriers group. Providing email
notifications is great and all, but how about an API. Let end users and
interested parties build clients for you like this officially, rather
than us having to hack-around just to work around the notion that
emails are good for automated systems....